requirements.txt
asgiref
autopep8==2.0.2
aiogram==2.16.0
beautifulsoup4==4.12.2
dj-database-url==0.5.0
Django==2.2.10
django-admin-atlantis==1.0.0
django-admin-material-dashboard==1.0.10
django-admin-soft-dashboard==1.0.12
django-crispy-forms==2.0
python-telegram-bot==13.7
requests==2.25.1
pycodestyle
pytz
sqlparse
Unipath
dj-database-url
python-decouple
gunicorn
whitenoise